Extract selectors and fix sources display
